Bailed on ssr, trying html.js route
Overidding html.js to inject custom Optimole JS script.
